# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SquadGod API — futuristic football AI platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  position: string;
  rating: number;
  pace: number;
  shooting: number;
  passing: number;
  dribbling: number;
  defending: number;
  physical: number;
  nationality: string;
  club: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface SquadMember {
  id: number;
  playerId: number;
  positionSlot: string;
  player: Player;
}

export interface Squad {
  id: number;
  name: string;
  formation: string;
  overallRating: number;
  players: SquadMember[];
  createdAt: string;
}

export interface SquadInput {
  /** @minLength 1 */
  name: string;
  formation: string;
}

export interface SquadUpdate {
  name?: string;
  formation?: string;
}

export interface SquadPlayerInput {
  playerId: number;
  positionSlot: string;
}

export interface SquadsOverview {
  totalSquads: number;
  totalPlayers: number;
  avgOverallRating: number;
  topFormation: string;
}

export interface SquadAnalysis {
  squadId: number;
  strengths: string[];
  weaknesses: string[];
  suggestions: string[];
  overallGrade: string;
}

export interface MatchPrediction {
  id: number;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  confidence: number;
  aiAnalysis: string;
  createdAt: string;
}

export interface MatchInput {
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  homeSquadId?: number | null;
  /** @nullable */
  awaySquadId?: number | null;
}

export type ListPlayersParams = {
search?: string;
position?: string;
minRating?: number;
};

export type ListTopRatedPlayersParams = {
limit?: number;
};