# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SquadGod API — futuristic football AI platform
 * OpenAPI spec version: 0.1.0
 */

export interface MatchInput {
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  homeSquadId?: number | null;
  /** @nullable */
  awaySquadId?: number | null;
}