# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SquadGod API — futuristic football AI platform
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  position: string;
  rating: number;
  pace: number;
  shooting: number;
  passing: number;
  dribbling: number;
  defending: number;
  physical: number;
  nationality: string;
  club: string;
  /** @nullable */
  imageUrl?: string | null;
}