# lib/api-zod/src/generated/types/squad.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SquadGod API — futuristic football AI platform
 * OpenAPI spec version: 0.1.0
 */
import type { SquadMember } from './squadMember';

export interface Squad {
  id: number;
  name: string;
  formation: string;
  overallRating: number;
  players: SquadMember[];
  createdAt: string;
}