# lib/api-zod/src/generated/types/squadMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SquadGod API — futuristic football AI platform
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';

export interface SquadMember {
  id: number;
  playerId: number;
  positionSlot: string;
  player: Player;
}