# lib/api-zod/src/generated/types/squadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SquadGod API — futuristic football AI platform
 * OpenAPI spec version: 0.1.0
 */

export interface SquadUpdate {
  name?: string;
  formation?: string;
}